fix hook order race with telemetry (#930)
Fixes hook race condition for telemetry (onactivate, onconfigchanged)

mattermost-server and pluginapi deps have been updated, opengraph forced as an indirect dep to avoid a ambiguous dep conflict
