+++ go.mod
@@ -9,13 +9,15 @@ require (
 	github.com/dgrijalva/jwt-go v3.2.0+incompatible
 	github.com/gorilla/mux v1.8.0
 	github.com/jarcoal/httpmock v1.0.8
-	github.com/mattermost/mattermost-plugin-api v0.0.27
+	github.com/mattermost/mattermost-plugin-api v0.1.3
 	github.com/mattermost/mattermost-plugin-autolink v1.2.2-0.20210709183311-c8fa30db649f
-	github.com/mattermost/mattermost-server/v6 v6.5.2
+	github.com/mattermost/mattermost-server/v6 v6.0.0-20221109191448-21aec2741bfe
 	github.com/mholt/archiver/v3 v3.5.1
 	github.com/pkg/errors v0.9.1
 	github.com/rbriski/atlassian-jwt v0.0.0-20180307182949-7bb4ae273058
-	github.com/stretchr/testify v1.7.0
+	github.com/stretchr/testify v1.8.0
 	golang.org/x/oauth2 v0.0.0-20211104180415-d3ed0bb246c8
 	golang.org/x/text v0.3.7
 )
+
+exclude github.com/dyatlov/go-opengraph v0.0.0-20210112100619-dae8665a5b09
